Add Tuple Wrapper (in Ruby (-> in your face))
